added login check, and redirect on successful login, or message if failed login
